[regression] split template_2 into two test cases
