+++ regression/esbmc-cpp/template/template_3/test.desc
@@ -9,5 +9,5 @@
   <item_07_test_importance>High</item_07_test_importance>
   <item_08_execution_type>Automated</item_08_execution_type>
   <item_09_author>Ceteli/UFAM</item_09_author>
-<item_10_mode>CORE</item_10_mode>
+<item_10_mode>KNOWNBUG</item_10_mode>
 </test-case>
