+++ regression/esbmc-cpp/template/template_3_bug/test.desc
@@ -5,7 +5,7 @@
   <item_03_summary>Aims to analyze the main function from module: esbmc-list</item_03_summary>
   <item_04_file_C_to_test>main.cpp </item_04_file_C_to_test>
   <item_05_option_to_run_esbmc>--unwind 10 --no-unwinding-assertions -I /__w/esbmc/esbmc/src/cpp/library/ --memlimit 14000000 --timeout 900</item_05_option_to_run_esbmc>
-  <item_06_expected_result>^VERIFICATION SUCCESSFUL$</item_06_expected_result>
+  <item_06_expected_result>^VERIFICATION FAILED$</item_06_expected_result>
   <item_07_test_importance>High</item_07_test_importance>
   <item_08_execution_type>Automated</item_08_execution_type>
   <item_09_author>Ceteli/UFAM</item_09_author>
