+++ regression/esbmc-cpp/template/template_3_bug/testllvm.desc
@@ -1,13 +1,12 @@
-<?xml version='1.0' encoding='utf-8'?>
+<?xml version='1.0' encoding='UTF-8'?>
 <test-case>
   <item_01_test_case_version>Versão 1.0</item_01_test_case_version>
   <item_02_id>esbmc-list</item_02_id>
   <item_03_summary>Aims to analyze the main function from module: esbmc-list</item_03_summary>
-  <item_04_file_C_to_test>main.cpp </item_04_file_C_to_test>
-  <item_05_option_to_run_esbmc>--unwind 10 --no-unwinding-assertions -I /__w/esbmc/esbmc/src/cpp/library/ --memlimit 14000000 --timeout 900</item_05_option_to_run_esbmc>
-  <item_06_expected_result>^VERIFICATION SUCCESSFUL$</item_06_expected_result>
+  <item_04_file_C_to_test>main.bc</item_04_file_C_to_test>
+  <item_05_option_to_run_esbmc>--ignore-missing-function-bodies --max-loop-iterations=10 --no-max-loop-iterations-checks</item_05_option_to_run_esbmc>
+  <item_06_expected_result>^VERIFICATION FAILED$</item_06_expected_result>
   <item_07_test_importance>High</item_07_test_importance>
   <item_08_execution_type>Automated</item_08_execution_type>
   <item_09_author>Ceteli/UFAM</item_09_author>
-<item_10_mode>CORE</item_10_mode>
 </test-case>
